+++ components/utilities/updateAuctions.js
@@ -1,8 +1,16 @@
 let fs = require("fs");
 let Polkadot = require("@polkadot/api");
 
+let AuctionVariables = require("./auctionVariables");
+const PolkadotSlotLeasePeriod = AuctionVariables.PolkadotSlotLeasePeriod;
+const PolkadotSlotLeaseOffset = AuctionVariables.PolkadotSlotLeaseOffset;
+const PolkadotLeasePeriodPerSlot = AuctionVariables.PolkadotLeasePeriodPerSlot;
+const KusamaSlotLeasePeriod = AuctionVariables.KusamaSlotLeasePeriod;
+const KusamaSlotLeaseOffset = AuctionVariables.KusamaSlotLeaseOffset;
+const KusamaLeasePeriodPerSlot = AuctionVariables.KusamaLeasePeriodPerSlot;
+const FutureBlock = AuctionVariables.FutureBlock;
+
 let API = undefined;
-const FutureBlock = "0x0000000000000000000000000000000000000000000000000000000000000000";
 
 LoadAPI().then(() => {
 	fs.readFile("./components/utilities/data/PolkadotAuctions.json", "utf8", async function readFileCallback(err, data) {
@@ -34,7 +42,13 @@ LoadAPI().then(() => {
 
 			// Write
 			json = JSON.stringify(existingAuctions); //convert it back to json
-			fs.writeFile("auctions-sample.json", json, "utf8", callback);
+			fs.writeFile("auctions-sample.json", json, "utf8", async function writeFileCallback(err) {
+				if (err) {
+					console.log(err);
+				} else {
+					console.log("Update Complete.");
+				}
+			});
 		}
 	})
 }
@@ -45,6 +59,37 @@ async function LoadAPI() {
 	API = await Polkadot.ApiPromise.create({ provider: WSProvider });
 }
 
-function callback() {
-  console.log("Done.");
+// Get the auction end, on-board start and end blocks from auction start block
+async function GetAuctionBlocks(api, startBlock, chain) {
+	const hash = await BlockToHash(startBlock);
+	if (hash !== FutureBlock) {
+		const apiAt = await api.at(hash);
+		const [auctionLeasePeriod, auctionEndBlock] = (await apiAt.query.auctions.auctionInfo()).toJSON();
+		if (chain === "Polkadot") {
+			const onboardStartBlock = auctionLeasePeriod * PolkadotSlotLeasePeriod + PolkadotSlotLeaseOffset;
+			const onboardEndBlock = onboardStartBlock + DaysToBlocks(PolkadotLeasePeriodPerSlot * 12 * 7);
+			return [auctionEndBlock, onboardStartBlock, onboardEndBlock]
+		}
+		else if (chain === "Kusama") {
+			const onboardStartBlock = auctionLeasePeriod * KusamaSlotLeasePeriod + KusamaSlotLeaseOffset;
+			const onboardEndBlock = onboardStartBlock + DaysToBlocks(KusamaLeasePeriodPerSlot * 6 * 7);
+			return [auctionEndBlock, onboardStartBlock, onboardEndBlock]
+		}
+	}
+	else {
+		// We are dealing with future blocks - TODO use subscan instead of PolkadotJS?
+		return [0, 0, 0];
+	}
+}
+
+// Block number to block hash
+async function BlockToHash(api, block) {
+	const hash = await api.rpc.chain.getBlockHash(block);
+	return hash;
+}
+
+// Convert an integer representing number of days block count for that time span
+function DaysToBlocks(days) {
+	const blocks = (days / 6) * 86400;
+	return blocks;
 }
